feat(spotify): unify tabs-on-active model across all three layouts
Generalize the Default layout's "tabs appear when the widget is
selected" model into one shared PersonalSpotifyAdaptiveLayout used by
all three personal-Spotify variants (default / minimal / small), so they
behave consistently. The only per-variant difference is the player
surface: default uses the Now Playing card, minimal uses the full-bleed
art view, small uses the compact bar.

At rest the widget shows only the player surface. When selected, the
Songs/Playlists/search tab bar appears — as a top strip while browsing
(all variants) and over the player for default/small, while minimal
overlays the pills on the art with a top scrim. Selecting a track plays
it and returns to the player view on every variant.

Removes the old browseOpen tap-to-open overlay and the onOpenBrowse prop
from the compact/minimal surfaces (now pure player surfaces). Deletes the
now-unused PersonalSpotifyBrowsePanel and PersonalSpotifyTabs (and their
tests) after confirming zero remaining references.
