+++ components/widgets/MusicWidget/PersonalSpotifyBrowser.tsx
@@ -1,16 +1,12 @@
-import React, { useCallback, useState } from 'react';
+import React, { useCallback } from 'react';
 import { WidgetData, MusicConfig } from '@/types';
 import { WidgetLayout } from '@/components/widgets/WidgetLayout';
 import { useDashboard } from '@/context/useDashboard';
 import { useSpotifyAuth } from '@/hooks/useSpotifyAuth';
 import { useSpotifyWebPlayback } from '@/hooks/useSpotifyWebPlayback';
 import { playOnDevice } from '@/utils/spotifyAuth';
-import { SpotifyBrowserTab } from './PersonalSpotifyTabs';
 import { SpotifyPlayablePick } from './PersonalSpotifyLibraryTab';
-import { PersonalSpotifyBrowsePanel } from './PersonalSpotifyBrowsePanel';
-import { PersonalSpotifyCompactBar } from './PersonalSpotifyCompactBar';
-import { PersonalSpotifyMinimalView } from './PersonalSpotifyMinimalView';
-import { PersonalSpotifyDefaultLayout } from './PersonalSpotifyDefaultLayout';
+import { PersonalSpotifyAdaptiveLayout } from './PersonalSpotifyAdaptiveLayout';
 
 interface Props {
   widget: WidgetData;
@@ -20,31 +16,24 @@ export const PersonalSpotifyBrowser: React.FC<Props> = ({ widget }) => {
   const config = widget.config as MusicConfig;
   const { layout = 'default' } = config;
   const { updateWidget, selectedWidgetId } = useDashboard();
-  // The Default layout reveals its tab strip only while the widget is the
-  // selected one — the same flag that drives the widget toolbar's visibility.
+  // The tab strip is revealed only while the widget is the selected one — the
+  // same flag that drives the widget toolbar's visibility. This now applies to
+  // ALL THREE layouts (default / minimal / small) via the shared adaptive
+  // layout below.
   const isActive = selectedWidgetId === widget.id;
   // Note: useSpotifyAuth exposes `connect` and `disconnect` but no `reconnect`.
   // handleReconnect below calls disconnect() then connect() to achieve a
-  // reconnect flow. The `reconnect` name is preserved in the prop passed to
-  // PersonalSpotifyLibraryTab for clarity.
+  // reconnect flow.
   const { isPremium, getAccessToken, disconnect, connect } = useSpotifyAuth();
-
-  const [activeTab, setActiveTab] = useState<SpotifyBrowserTab>('library');
-  // Small/minimal layouts have no inline tabs, so tapping the surface opens
-  // the full browse UI as a temporary full-cover overlay.
-  const [browseOpen, setBrowseOpen] = useState(false);
 
   const currentUri = config.personalSpotifyUrl ?? null;
 
   // The Web Playback SDK is owned here (above the tabs) so the playback device
-  // survives tab switches — a track tapped in Library plays even before the
-  // Now Playing tab is opened, and leaving Now Playing doesn't stop the music.
+  // survives view switches — a track tapped in a list plays even before the
+  // player view is shown, and leaving the player doesn't stop the music.
   // Disabled for Free accounts (SDK requires Premium); they get the embed.
   // currentUri is the target the device starts on first play (reload-resume).
   const playback = useSpotifyWebPlayback(isPremium, getAccessToken, currentUri);
-
-  const isAudioActive =
-    playback.isPlaying || (!isPremium && Boolean(currentUri));
 
   const handleReconnect = useCallback(async () => {
     await disconnect();
@@ -56,11 +45,8 @@ export const PersonalSpotifyBrowser: React.FC<Props> = ({ widget }) => {
       updateWidget(widget.id, {
         config: { personalSpotifyUrl: pick.uri },
       });
-      // The Default layout owns its own view state (PersonalSpotifyDefaultLayout
-      // returns to the player after a row tap), so nothing to do here for it.
-      // Compact/minimal layouts have no tabs — selecting a track in their browse
-      // overlay closes it.
-      if (layout !== 'default') setBrowseOpen(false);
+      // The adaptive layout owns its own view state (returns to the player
+      // after a row tap), so nothing to do here for navigation.
       if (!isPremium) return;
       const token = await getAccessToken();
       if (!token || !playback.deviceId) return;
@@ -72,17 +58,10 @@ export const PersonalSpotifyBrowser: React.FC<Props> = ({ widget }) => {
         console.warn('[PersonalSpotifyBrowser.handlePlay] play failed', err);
       }
     },
-    [
-      updateWidget,
-      widget.id,
-      layout,
-      isPremium,
-      getAccessToken,
-      playback.deviceId,
-    ]
+    [updateWidget, widget.id, isPremium, getAccessToken, playback.deviceId]
   );
 
-  // Shared playback props handed to the leaf "now playing" surfaces.
+  // Shared playback props handed to the leaf player surfaces.
   const playbackProps = {
     url: currentUri,
     thumbnail: config.personalSpotifyThumbnail,
@@ -101,76 +80,16 @@ export const PersonalSpotifyBrowser: React.FC<Props> = ({ widget }) => {
     onToggleShuffle: () => void playback.toggleShuffle(),
   };
 
-  // The browse panel (tab strip + active-tab body), shared between the inline
-  // default layout and the small/minimal overlay. `onClose` is only passed in
-  // overlay mode (renders the X button).
-  const renderBrowsePanel = (onClose?: () => void) => (
-    <PersonalSpotifyBrowsePanel
-      activeTab={activeTab}
-      onTabChange={setActiveTab}
-      isAudioActive={isAudioActive}
-      currentUri={currentUri}
-      onPlay={handlePlay}
-      onReconnect={handleReconnect}
-      onSwitchToLibrary={() => setActiveTab('library')}
-      playbackProps={playbackProps}
-      onClose={onClose}
-    />
-  );
-
-  // Small layout → single now-playing strip (no tabs). Tapping the art/title
-  // opens the browse overlay so a new track can be picked.
-  if (layout === 'small') {
-    return (
-      <WidgetLayout
-        padding="p-0"
-        content={
-          <div className="w-full h-full relative">
-            <PersonalSpotifyCompactBar
-              {...playbackProps}
-              onOpenBrowse={() => setBrowseOpen(true)}
-            />
-            {browseOpen && (
-              <div className="absolute inset-0 z-20 bg-slate-900">
-                {renderBrowsePanel(() => setBrowseOpen(false))}
-              </div>
-            )}
-          </div>
-        }
-      />
-    );
-  }
-
-  // Minimal layout → full-bleed artwork + centered play (no tabs). Tapping the
-  // artwork opens the browse overlay.
-  if (layout === 'minimal') {
-    return (
-      <WidgetLayout
-        padding="p-0"
-        content={
-          <div className="w-full h-full relative">
-            <PersonalSpotifyMinimalView
-              {...playbackProps}
-              onOpenBrowse={() => setBrowseOpen(true)}
-            />
-            {browseOpen && (
-              <div className="absolute inset-0 z-20 bg-slate-900">
-                {renderBrowsePanel(() => setBrowseOpen(false))}
-              </div>
-            )}
-          </div>
-        }
-      />
-    );
-  }
-
-  // Default layout → player at rest; reveals Songs/Playlists/search tabs only
-  // while the widget is selected. View state is owned by the layout component.
+  // All three layouts route through the shared adaptive layout. At rest it
+  // shows only the variant's player surface; when the widget is selected the
+  // Songs/Playlists/search tab bar appears (overlaid on minimal, top strip on
+  // default/small).
   return (
     <WidgetLayout
       padding="p-0"
       content={
-        <PersonalSpotifyDefaultLayout
+        <PersonalSpotifyAdaptiveLayout
+          variant={layout}
           isActive={isActive}
           currentUri={currentUri}
           onPlay={handlePlay}
